backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
greenlet==3.0.3
httpx==0.25.2
python-dotenv==1.0.0
PyPDF2==3.0.1
GitPython==3.1.40
apscheduler==3.10.4
email-validator==2.1.0
